+++ debug.c
@@ -11,11 +11,12 @@ void printbuffer(unsigned char* buffer, size_t len)
         printf("%04X: ", (unsigned int)offset);
         for(i = 0; i < 8 && i+offset < len; i++)
         {
-            printf("%02X ", buffer[i+offset]);
+            printf("%02X ", (unsigned int)buffer[i+offset]);
         }
         for(i = 0; i < 8 && i+offset < len; i++)
         {
-            if(buffer[i+offset] < 0x20 || buffer[i+offset] > 0x7E)
+            if((unsigned int)buffer[i+offset] < 0x20 || 
+               (unsigned int)buffer[i+offset] > 0x7E)
             {
                 printf(".");
             }
